+++ test/PHPUnit/model/EntiteSQLTest.php
@@ -94,16 +94,6 @@ class EntiteSQLTest extends PastellTestCase
         $this->assertEquals(1, $id_e);
     }
 
-    public function testGetCollectiviteAncetre()
-    {
-        $this->assertEquals(1, $this->entiteSQL->getCollectiviteAncetre(1));
-    }
-
-    public function testGetCollectiviteAncetreService()
-    {
-        $this->assertEquals(1, $this->entiteSQL->getCollectiviteAncetre(2));
-    }
-
     public function testGetFilleInfoNavigation()
     {
         $this->assertEquals(