+++ CHANGELOG.md
@@ -65,6 +65,8 @@
   d'initialisation du conteneur.
 - Suppression du modèle `Document`, remplacé par `DocumentSQL`
 - Suppression du modèle `Utilisateur`, remplacé par `UtilisateurSQL`
+- Suppression du modèle `Entite`, remplacé par `EntiteSQL`
+  (suppression du type d'entité "service" et des classes dépréciées AccuserReception, Envoyer)
 - Suppression des fonctions :
   - SAEConnecteur::getAcuseReception
   - SAEConnecteur::getReply